arch/dumponexit: unify dump on exit to common code
remove arch implement and move to common code
