+++ sched/misc/Make.defs
@@ -24,6 +24,10 @@ ifeq ($(CONFIG_ARCH_DEADLOCKDUMP),y)
 CSRCS += deadlock.c
 endif
 
+ifeq ($(CONFIG_DUMP_ON_EXIT),y)
+CSRCS += dump.c
+endif
+
 # Include init build support
 
 DEPPATH += --dep-path misc
